melita/requirements.txt
anyascii==0.3.1
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.11.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.7
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.6.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
l18n==2021.3
mypy-extensions==1.0.0
openpyxl==3.1.1
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.0
psycopg2==2.9.5
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
telepath==0.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
wagtail==4.2
webencodings==0.5.1
Willow==1.4.1
